+++ Risk.Engine/Class1.cs
@@ -1,6 +1,0 @@
-﻿namespace Risk.Engine;
-
-public class Class1
-{
-
-}